requirements.txt
soundcard
numpy
spotipy
pydub
mutagen
customtkinter
pywinauto
Pillow
python-dotenv
